// src/types/API.ts
export type NavHistoryInfo = {
  __typename: "NavHistoryInfo";
  totalCount?: number | null;
  poolSlotInfoId?: number | null;
  symbol?: string | null;
  currencyDecimals?: number | null;
  allTimeHigh?: string | null;
  serialData?: Array<NavHistoryData | null> | null;
};

export type NavHistoryData = {
  __typename: "NavHistoryData";
  nav?: string | null;
  adjustedNav?: string | null;
  navDate?: string | null;
};